TAS: move expecatations out for reuse
